+++ Dracula/ModEntry.cs
@@ -116,9 +116,12 @@ public sealed class ModEntry : SimpleMod
 	];
 
 	internal static IReadOnlyList<Type> CommonArtifacts { get; } = [
+		typeof(LacerationsArtifact),
 		typeof(MasochismArtifact),
 		typeof(ThinBloodArtifact),
 		typeof(WingsOfNightArtifact),
+		
+		// old
 		typeof(TheCountArtifact),
 	];